+++ codex-rs/tui2/src/lib.rs
@@ -58,9 +58,7 @@ pub mod live_wrap;
 mod markdown;
 mod markdown_render;
 mod markdown_stream;
-#[cfg(test)]
 mod model_migration;
-mod model_migration_logic;
 mod notifications;
 pub mod onboarding;
 mod oss_selection;
